pkp/pkp-lib#8409 jobs folder and namespace update
