+++ classes/observers/listeners/BatchMetadataChangedListener.php
@@ -18,7 +18,7 @@ declare(strict_types=1);
 namespace PKP\observers\listeners;
 
 use Illuminate\Events\Dispatcher;
-use PKP\Jobs\Metadata\BatchMetadataChangedJob;
+use PKP\jobs\metadata\BatchMetadataChangedJob;
 
 use PKP\observers\events\BatchMetadataChanged;
