+++ classes/observers/listeners/RemoveSubmissionFileFromSearchIndex.php
@@ -18,7 +18,7 @@ declare(strict_types=1);
 namespace PKP\observers\listeners;
 
 use Illuminate\Events\Dispatcher;
-use PKP\Jobs\Submissions\RemoveSubmissionFileFromSearchIndexJob;
+use PKP\jobs\submissions\RemoveSubmissionFileFromSearchIndexJob;
 
 use PKP\observers\events\SubmissionFileDeleted;
