+++ classes/task/StatisticsReport.php
@@ -19,8 +19,8 @@ use APP\core\Application;
 use DateTimeImmutable;
 use Illuminate\Support\Facades\Bus;
 use PKP\db\DAORegistry;
-use PKP\Jobs\Notifications\StatisticsReportMail;
-use PKP\Jobs\Notifications\StatisticsReportNotify;
+use PKP\jobs\notifications\StatisticsReportMail;
+use PKP\jobs\notifications\StatisticsReportNotify;
 use PKP\mail\Mailer;
 use PKP\notification\managerDelegate\EditorialReportNotificationManager;
 use PKP\notification\NotificationSubscriptionSettingsDAO;
