+++ tools/jobs.php
@@ -29,8 +29,8 @@ use Illuminate\Console\OutputStyle;
 use PKP\cliTool\CommandLineTool;
 use PKP\config\Config;
 use PKP\job\models\Job as PKPJobModel;
-use PKP\Jobs\TestJobs\TestJobFailure;
-use PKP\Jobs\TestJobs\TestJobSuccess;
+use PKP\jobs\testJobs\TestJobFailure;
+use PKP\jobs\testJobs\TestJobSuccess;
 use PKP\queue\WorkerConfiguration;
 use Symfony\Component\Console\Exception\CommandNotFoundException;
 use Symfony\Component\Console\Exception\InvalidArgumentException as CommandInvalidArgumentException;
